+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - The `Names` struct is no longer an iterator, but a container.
   You have to call `into_iter()` manually.
 - Most of the `Name` struct methods become public fields.
+- `Face::units_per_em` no longer returns `Option`.
 
 ### Removed
 - (c-api) `ttfp_glyph_class`, `ttfp_get_glyph_class`, `ttfp_get_glyph_mark_attachment_class`,